+++ repo/repo-sql-impl/src/main/java/com/evolveum/midpoint/repo/sql/data/common/ROwnable.java
@@ -29,5 +29,5 @@ public interface ROwnable {
 
     String getOwnerOid();
 
-    Long getOwnerId();
+    Short getOwnerId();
 }
