+++ repo/repo-sql-impl/src/main/java/com/evolveum/midpoint/repo/sql/data/common/any/RAnyClob.java
@@ -38,7 +38,7 @@ public class RAnyClob implements RAnyValue {
     //owner entity
     private RAnyContainer anyContainer;
     private String ownerOid;
-    private Long ownerId;
+    private Short ownerId;
     private RContainerType ownerType;
 
     private boolean dynamic;
@@ -78,7 +78,7 @@ public class RAnyClob implements RAnyValue {
 
     @Id
     @Column(name = "anyContainer_owner_id")
-    public Long getOwnerId() {
+    public Short getOwnerId() {
         if (ownerId == null && anyContainer != null) {
             ownerId = anyContainer.getOwnerId();
         }
@@ -171,7 +171,7 @@ public class RAnyClob implements RAnyValue {
         this.ownerOid = ownerOid;
     }
 
-    public void setOwnerId(Long ownerId) {
+    public void setOwnerId(Short ownerId) {
         this.ownerId = ownerId;
     }
