+++ repo/repo-sql-impl/src/main/java/com/evolveum/midpoint/repo/sql/data/common/any/RAnyString.java
@@ -36,7 +36,7 @@ public class RAnyString implements RAnyValue {
     //owner entity
     private RAnyContainer anyContainer;
     private String ownerOid;
-    private Long ownerId;
+    private Short ownerId;
     private RContainerType ownerType;
 
     private boolean dynamic;
@@ -76,7 +76,7 @@ public class RAnyString implements RAnyValue {
 
     @Id
     @Column(name = "anyContainer_owner_id")
-    public Long getOwnerId() {
+    public Short getOwnerId() {
         if (ownerId == null && anyContainer != null) {
             ownerId = anyContainer.getOwnerId();
         }
@@ -151,7 +151,7 @@ public class RAnyString implements RAnyValue {
         this.ownerOid = ownerOid;
     }
 
-    public void setOwnerId(Long ownerId) {
+    public void setOwnerId(Short ownerId) {
         this.ownerId = ownerId;
     }
